LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/srcdoc/srcdoc-history-entries-expected.txt


FAIL srcdoc history entries: the iframe itself navigates assert_equals: expected 2 but got 1
FAIL srcdoc history entries: the container window navigates assert_equals: expected 2 but got 1
